Replay text through one DrawTarget-generic function (#578) (#669)
The raster, SVG, PDF and BoundsTarget replays of the deferred label
list collapse into one `_replay_text_requests[T: DrawTarget]`, calling
the trait's draw_text, draw_text_runs and draw_line_aa. Raster and PDF
output is byte-identical; SVG anchors now print at the backend's fixed
precision (`x="140.000"`), which is the same number and the same
picture, so the tests that pinned that markup follow the backend and
the output digest's SVG columns move.

+++ tests/test_color_scales.mojo
@@ -103,7 +103,7 @@ def _text_y(svg: String, label: String) raises -> Int:
         raise Error("no y attribute on the <text> for " + label)
     var v = k + key.byte_length()
     var end = element.find('"', v)
-    return Int(String(element[byte=v:end]))
+    return Int(Float64(String(element[byte=v:end])))
 
 
 def _assert_same_color(got: Color, want: Color, label: String) raises:
